+++ pico_esp8266_at/src/gblink.c
@@ -1,0 +1,17 @@
+#include "gblink.h"
+
+void trigger_spi(spi_inst_t *spi, uint baudrate) {
+    //spi_init
+    reset_block(spi == spi0 ? RESETS_RESET_SPI0_BITS : RESETS_RESET_SPI1_BITS);
+    unreset_block_wait(spi == spi0 ? RESETS_RESET_SPI0_BITS : RESETS_RESET_SPI1_BITS);
+
+    spi_set_baudrate(spi, baudrate);    
+    spi_set_format(spi, 8, SPI_CPOL_1, SPI_CPOL_1, SPI_MSB_FIRST);
+    hw_set_bits(&spi_get_hw(spi)->dmacr, SPI_SSPDMACR_TXDMAE_BITS | SPI_SSPDMACR_RXDMAE_BITS);
+    spi_set_slave(spi, true);
+    
+    // Set the first data into the buffer
+    spi_get_hw(spi)->dr = 0xD2;
+
+    hw_set_bits(&spi_get_hw(spi)->cr1, SPI_SSPCR1_SSE_BITS);
+}
